requirements.txt
Flask>=2.2.2
marshmallow==3.7.1
requests==2.24.0
flake8==3.8.3
coverage==5.2.1
pytest==6.0.1
cryptography==3.3.2
Authlib==0.15.4
gunicorn==20.0.4
PyJWT==2.1.0
google-api-python-client
